django/room_allocation/requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
